Split wbi_core into exceptions and functions
