+++ test/test_all.py
@@ -4,7 +4,7 @@ import unittest
 
 import requests
 
-from wikibaseintegrator import wbi_core, wbi_fastrun
+from wikibaseintegrator import wbi_core, wbi_fastrun, wbi_functions
 from wikibaseintegrator.wbi_core import MWApiError
 
 __author__ = 'Sebastian Burgstaller-Muehlbacher'
@@ -14,13 +14,13 @@ __license__ = 'AGPLv3'
 class TestMediawikiApiCall(unittest.TestCase):
     def test_all(self):
         with self.assertRaises(MWApiError):
-            wbi_core.FunctionsEngine.mediawiki_api_call_helper(data={'format': 'json', 'action': 'wbgetentities', 'ids': 'Q42'}, mediawiki_api_url="http://www.wikidataaaaaaa.org",
-                                                               max_retries=3, retry_after=1, allow_anonymous=True)
+            wbi_functions.mediawiki_api_call_helper(data={'format': 'json', 'action': 'wbgetentities', 'ids': 'Q42'}, mediawiki_api_url="http://www.wikidataaaaaaa.org",
+                                                    max_retries=3, retry_after=1, allow_anonymous=True)
         with self.assertRaises(requests.HTTPError):
-            wbi_core.FunctionsEngine.mediawiki_api_call_helper(data=None, mediawiki_api_url="http://httpbin.org/status/400", max_retries=3, retry_after=1, allow_anonymous=True)
+            wbi_functions.mediawiki_api_call_helper(data=None, mediawiki_api_url="http://httpbin.org/status/400", max_retries=3, retry_after=1, allow_anonymous=True)
 
-        test = wbi_core.FunctionsEngine.mediawiki_api_call_helper(data={'format': 'json', 'action': 'wbgetentities', 'ids': 'Q42'}, max_retries=3, retry_after=1,
-                                                                  allow_anonymous=True)
+        test = wbi_functions.mediawiki_api_call_helper(data={'format': 'json', 'action': 'wbgetentities', 'ids': 'Q42'}, max_retries=3, retry_after=1,
+                                                       allow_anonymous=True)
         print(test)
