+++ test/test_wbi_login.py
@@ -5,7 +5,7 @@ import sys
 
 import pytest
 
-from wikibaseintegrator import wbi_login, wbi_core
+from wikibaseintegrator import wbi_login, wbi_functions
 
 # look for environment variables. if none set, don't do anything
 WDUSER = os.getenv("WDUSER")
@@ -23,4 +23,4 @@ def test_write():
     if WDUSER and WDPASS:
         login = wbi_login.Login(WDUSER, WDPASS)
         with pytest.raises(ValueError):
-            wbi_core.FunctionsEngine.mediawiki_api_call_helper(data=None, login=login, mediawiki_api_url='http://unsdfdskfjljzkerezr.org/w/api.php')
+            wbi_functions.mediawiki_api_call_helper(data=None, login=login, mediawiki_api_url='http://unsdfdskfjljzkerezr.org/w/api.php')
